+++ omnigibson/utils/motion_planning_utils.py
@@ -451,7 +451,7 @@ def set_base_and_detect_collision(context, pose):
     translation = lazy.pxr.Gf.Vec3d(*th.tensor(pose[0], dtype=th.float32).tolist())
     robot_copy.prims[robot_copy_type].GetAttribute("xformOp:translate").Set(translation)
 
-    orientation = th.tensor(pose[1], dtype=float)[[3, 0, 1, 2]]
+    orientation = th.tensor(pose[1], dtype=th.float32)[[3, 0, 1, 2]]
     robot_copy.prims[robot_copy_type].GetAttribute("xformOp:orient").Set(lazy.pxr.Gf.Quatd(*orientation.tolist()))
 
     return detect_robot_collision(context)
@@ -482,7 +482,7 @@ def set_arm_and_detect_collision(context, joint_pos):
                 mesh_pose = T.pose_transform(*pose, *relative_pose)
                 translation = lazy.pxr.Gf.Vec3d(*th.tensor(mesh_pose[0], dtype=th.float32).tolist())
                 mesh.GetAttribute("xformOp:translate").Set(translation)
-                orientation = th.tensor(mesh_pose[1], dtype=float)[[3, 0, 1, 2]]
+                orientation = th.tensor(mesh_pose[1], dtype=th.float32)[[3, 0, 1, 2]]
                 mesh.GetAttribute("xformOp:orient").Set(lazy.pxr.Gf.Quatd(*orientation.tolist()))
 
     return detect_robot_collision(context)
